feat: change location of dtos and separate them
