+++ Team0418Pokemon/Data/DTO/PokemonList/PokemonListDTO.swift
