+++ Team0418Pokemon/Data/DTO/PokemonSpecies/PokemonSpeciesDTO.swift
